refactor: Remove InfomaniakNetworkLogin from core
