+++ Package.swift
@@ -17,6 +17,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/Infomaniak/ios-dependency-injection", .upToNextMajor(from: "2.0.0")),
+        .package(url: "https://github.com/Infomaniak/ios-login", branch: "login-refactor"),
         .package(url: "https://github.com/Alamofire/Alamofire", .upToNextMajor(from: "5.8.0")),
         .package(url: "https://github.com/getsentry/sentry-cocoa", .upToNextMajor(from: "8.18.0")),
         .package(url: "https://github.com/realm/realm-swift", .upToNextMajor(from: "10.45.0")),
@@ -29,6 +30,7 @@ let package = Package(
             dependencies: [
                 "Alamofire",
                 .product(name: "InfomaniakDI", package: "ios-dependency-injection"),
+                .product(name: "InfomaniakLogin", package: "ios-login"),
                 .product(name: "Sentry", package: "sentry-cocoa"),
                 .product(name: "RealmSwift", package: "realm-swift"),
                 .product(name: "CocoaLumberjackSwift", package: "CocoaLumberjack"),
